iterate the config runner in prep for separation
