+++ .gitignore
@@ -2,7 +2,7 @@
 __pycache__/
 *.py[cod]
 *$py.class
-
+scripts/
 # Data files and imagery
 *.pkl
 *.png
